requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
google-generativeai==0.8.3
redis==5.0.8
pydantic-settings==2.4.0
python-dotenv==1.0.1
httpx==0.27.2